Everything functional changing and pending listed
[Fix] A SQLiteConnection object for database was leaked! Please fix your application to end transactions in progress properly and to close the database when it is no longer needed.
[Add] Delete method for user profile and associated with transactions data.
[Add] Make focusable while adding new or edit user profile data.
[Changed] All integer typed id to long.

[Pending] Backup to google drive.
[Pending] Restore from google drive.
[Pending] Display message for last sync backup or restore from google drive.
[Pending] Change theme light to dark vice versa.
[Pending] Add Settings acitivity.
[Pending] Add Preference for change theme in Settings acitivity.
[Pending] Add Spinner for change currency seperator in Settings acitivity.
[Pending] Add Preference for clear caches in Settings acitivity.
[Pending] Add Preference for app version info in Settings acitivity.
[Pending] Add Preference for update app in Settings acitivity.
[Pending] Make better UI for NagivationView.
[Pending] Display message when last locally database backup using SharedPreferences.

+++ app/src/main/java/com/ahmer/accounting/model/Transactions.kt
@@ -2,7 +2,7 @@ package com.ahmer.accounting.model
 
 class Transactions {
 
-    var userId: Int = 0
+    var userId: Long = 0
 
     var transId: Long = 0
